requirements.txt
requests>=2.31.0
python-telegram-bot>=20.7
apscheduler>=3.10.4
python-dotenv>=1.0.0
Pillow>=10.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.3
psycopg2-binary>=2.9.9
